+++ trunk/ManagedDigitalImageProcessing/Filters/EdgeDetectors/SobelOperatorResult.cs
@@ -56,7 +56,7 @@ namespace ManagedDigitalImageProcessing.Filters.EdgeDetectors
         /// <value>
         /// The X data.
         /// </value>
-        public byte[] XData { get; set; }
+        public int[] XData { get; set; }
         
         /// <summary>
         /// Gets or sets the Y data.
@@ -64,6 +64,6 @@ namespace ManagedDigitalImageProcessing.Filters.EdgeDetectors
         /// <value>
         /// The Y data.
         /// </value>
-        public byte[] YData { get; set; }
+        public int[] YData { get; set; }
     }
 }
